- 1.5/2 Added World Menu (and fixes) -

+++ ConsoleAdventure/Content/Scripts/WorldEngine/Generate/WorldPropertiesGenerator.cs
@@ -29,7 +29,7 @@ namespace ConsoleAdventure.Content.Scripts.WorldEngine.Generate
             }
             set
             {
-                ConsoleAdventure.progressBar.Progress = (uint)value;
+                ConsoleAdventure.progressBar.Progress = value;
             }
         }
